# api/tsoa/swagger.json
{
	"components": {
...
		"schemas": {
...
			"User": {
				"properties": {
					"notification": {
						"type": "boolean"
					},
					"pass": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"name": {
						"type": "string",
						"nullable": true
					},
					"id": {
						"type": "string"
					}
				},
				"required": [
					"notification",
					"pass",
					"email",
					"name",
					"id"
				],
				"type": "object",
				"description": "Model User"
			}
		},
		"securitySchemes": {}
	},
	"info": {
		"title": ""
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/user": {
			"post": {
				"operationId": "CreateUser",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"User Handler"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
		"/login": {
			"post": {
				"operationId": "Login",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Login Handler"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/User"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}